requirements.txt
streamlit
obspy
pandas
numpy
scipy
plotly
google-genai
python-dotenv
watchdog
